ai-model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.20.0
numpy<2
opencv-python==4.8.1.78
Pillow==10.0.1
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
scikit-learn>=1.3.2
scipy>=1.11.4
